Adjust imports, move coupon routines to user related files
