+++ web/src/actions/seasons.js
@@ -1,5 +1,5 @@
 import { toast } from "react-toastify";
-import { dispatchFireworks } from "../utils/fireworks-dispatcher";
+import dispatchFireworks from "../utils/fireworks-dispatcher";
 
 import {
   GET_ALL_SEASONS_SUCCESS,
@@ -25,8 +25,6 @@ import {
   SET_ACTIVE_TEAM,
   CREATE_TEAM_SUCCESS,
   CREATE_TEAM_FAILED,
-  VALIDATE_COUPON_SUCCESS,
-  VALIDATE_COUPON_FAILED,
 } from "../constants/actionTypes";
 
 import {
@@ -40,7 +38,6 @@ import {
   postValidateChallenge,
   postCloseChallenge,
   postCreateTeam,
-  postCouponValidation,
 } from "../api/seasons";
 
 //Season main actions
@@ -230,24 +227,5 @@ export const closeChallenge = subscriptionID => async dispatch => {
   } catch (error) {
     dispatch({ type: CLOSE_CHALLENGE_FAILED, payload: { error } });
     toast.error(`Close challenge ERROR!`);
-  }
-};
-
-//Coupon Actions
-export const fetchCouponValidation = (hash, teamID) => async dispatch => {
-  try {
-    const response = await postCouponValidation(hash, teamID);
-    dispatch({
-      type: VALIDATE_COUPON_SUCCESS,
-      payload: { team: response.data.coupon_validation.team },
-    });
-    toast.success(`Coupon validation success!`);
-    dispatchFireworks();
-  } catch (error) {
-    dispatch({
-      type: VALIDATE_COUPON_FAILED,
-      payload: { error: error.response },
-    });
-    toast.error(`Coupon validation error!`);
   }
 };
