+++ web/src/api/userSession.js
@@ -8,3 +8,9 @@ export function getUserSession() {
 export function deleteUserAccount(reason) {
   return baseApi.post(`/user/delete-account`, { reason: reason });
 }
+
+//Coupon calls
+
+export function postCouponValidation(hash, teamID) {
+  return baseApi.post(`/coupon-validation`, { hash: hash, team_id: teamID });
+}
